+++ CHANGELOG.md
@@ -7,9 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
-### Added
+### Changed
 
-- `resetOnMinMaxChange` to the `Slider` component.
+- Prevents `Slider` range from resetting when the min/max value changes.
 
 ## [9.146.3] - 2022-09-08