+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [9.146.4] - 2023-04-25
+
 ### Changed
 
 - Prevents `Slider` range from resetting when the min/max value changes.
@@ -4166,5 +4168,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - **Dropdown** and **Input** props `short` and `long`. Their widths should be defined by their parents.
 
 
-[Unreleased]: https://github.com/vtex/styleguide/compare/v9.146.1...HEAD
+[Unreleased]: https://github.com/vtex/styleguide/compare/v9.146.4...HEAD
 [9.146.1]: https://github.com/vtex/styleguide/compare/v9.146.0...v9.146.1
+
+[9.146.4]: https://github.com/vtex/styleguide/compare/v9.146.3...v9.146.4